// src/app/models/draw-tool-item.ts
export class DrawToolItem {
    Id?:any;
    ShapeType: string;
    SubType: string;
    Name: string;
    Description: string;
    FontSize: number;
    Color: string;
    StrokeThickness: number;
    BackColor: string;
    Opacity: number;
    ShapeGeo: string;
    HotSpotX: number;
    HotSpotY: number;
    Width: number;
    Height: number;
    UserId: string;
    Visible: number;
    Color3: string;
    IsDeleted: number;
    LineStyle: string;
    Radius: number;
    HTML_EditableLayerID?: any;
}